Added a Getting Started section to the Readme
This will serve as a basic introduction to the project for those who wish to understand the general/high-level basics of this project. It is not done yet, but I myself would've found this useful when I started working on this! Please let me know if there are sections I should add/delete